pyproject.toml
[project]
name = "helioscommand"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firecrawl>=1.0.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.9",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
]
